fix: flatten deployment script before uploading to enclave (#92)
This PR fixes an error when uploading local repos with a size that is
big enough. When trying to do so, it fails due to a limit of 100MB in
the Kurtosis engine. By flattening the contract before uploading, we
reduce this amount a lot. We could go lower but this should be enough
for now.