feat: add eigenda-proxy to EigenDA example (#149)
This PR adds the eigenda-proxy to the EigenDA example. In order to do
that, we had to change the points used from the inabox ones to the ones
taken from
https://github.com/Layr-Labs/rust-kzg-bn254/tree/master/tests/test-files

Doing this causes nodes to fail when dispersing blobs, so
`NODE_G2_POWER_OF_2_PATH` had to be removed.

Note that the proxy still fails due to an internal bug when batches
contain only a single blob.