fix: remove leading `0x` from private keys when whitelisting strategies (#171)
This PR removes a `0x` that we suffixed on the private key when
whitelisting strategies. It was causing the whitelisting to silently
fail because the private key already had a leading `0x`.
Also, it introduces an early return when whitelisting an empty list of
strategies, and adds `--confirmations 0` to each `cast send` call, to
reduce deployment time.