feat: add Aligned Layer's AVS example (#30)
This PR adds an example that deploys Aligned Layer and starts an
operator and aggregator. This isn't enabled in the CI due to images
taking too long to build.

What's left:
- add balance on the `ServiceManager` for the `BatcherPaymentService` at
startup. This is needed because the `ServiceManager` overcharges the
sender by about 2x, but always reimburses any extra. Note that this was
already added, but needs testing.